grid: sync / unsync OTA source in System Prefs
